Maven pre-integration work
