--- kotlin-examples/docker-compose.yml
@@ -1,12 +1,0 @@
-version: '3'
-services:
-  db:
-    image: couchbase/server:latest
-    ports:
-      - "8091:8091"
-      - "8092:8092"
-      - "8093:8093"
-      - "8094:8094"
-      - "11210:11210"
-    volumes:
-      - ./couchbase-setup.sh:/entrypoint.sh
