+++ kotlin-examples/src/integrationTest/kotlin/com/couchbase/client/kotlin/examples/util/ConnectionUtils.kt
@@ -1,0 +1,33 @@
+package com.couchbase.client.kotlin.examples.util
+
+import com.couchbase.client.kotlin.Bucket
+import com.couchbase.client.kotlin.Cluster
+import kotlinx.coroutines.runBlocking
+
+class ConnectionUtils {
+    companion object {
+        fun getCluster(): Cluster {
+            return Cluster.connect("couchbase://127.0.0.1", "Administrator", "password")
+        }
+
+        fun withCluster(block: suspend (Cluster) -> Unit) {
+            runBlocking {
+                val cluster = getCluster()
+                try {
+                    block(cluster)
+                } finally {
+                    cluster.disconnect()
+                }
+            }
+        }
+
+        fun withBucket(bucketName: String, block: suspend (Cluster, Bucket) -> Unit) {
+            withCluster { cluster ->
+                val bucket = cluster.bucket(bucketName)
+                block(cluster, bucket)
+            }
+        }
+
+    }
+
+}
