--- kotlin-examples/src/test/kotlin/com/couchbase/client/kotlin/examples/base/ConnectionBase.kt
@@ -1,7 +1,0 @@
-package com.couchbase.client.kotlin.examples.base
-
-class ConnectionBase {
-
-
-
-}
